+++ packages/client/src/app/ui/app.pcss
@@ -5,3 +5,7 @@
   padding: 0;
   box-sizing: border-box;
 }
+
+.ant-form-item-explain-error {
+  font-size: 14px;
+}
