+++ packages/client/src/app/ui/app.tsx
@@ -1,7 +1,7 @@
 import { themeConfig } from './theme-config'
 import { ConfigProvider, App as AntApp } from 'antd'
 import { RouterProvider } from 'react-router-dom'
-import { router } from './router'
+import { router } from './routing/router'
 import './app.pcss'
 
 export function App() {
