+++ packages/client/src/app/ui/routing/no-auth-routes.tsx
@@ -1,7 +1,8 @@
-import { useUserInfo } from '../../entities/user'
+import { useUserInfo } from '../../../entities/user'
 import { Spin } from 'antd'
 import { Navigate, Outlet } from 'react-router-dom'
-import { ROUTES } from '../../shared/config'
+import { ROUTES } from '../../../shared/config'
+import { AuthLayout } from '../layout/auth-layout'
 
 export function NoAuthRoutes() {
   const { data, isLoading } = useUserInfo()
@@ -14,5 +15,9 @@ export function NoAuthRoutes() {
     return <Navigate to={ROUTES.HOME} replace={true} />
   }
 
-  return <Outlet />
+  return (
+    <AuthLayout>
+      <Outlet />
+    </AuthLayout>
+  )
 }
