+++ packages/client/src/app/ui/routing/protected-routes.tsx
@@ -1,11 +1,11 @@
-import { useUserInfo } from '../../entities/user'
+import { useUserInfo } from '../../../entities/user'
 import { App, Button, Flex, Layout, Spin, Typography } from 'antd'
 import { Link, Navigate, Outlet, useLocation } from 'react-router-dom'
-import { ROUTES } from '../../shared/config'
+import { ROUTES } from '../../../shared/config'
 import { Content, Header } from 'antd/es/layout/layout'
 import React, { useEffect } from 'react'
-import { Logo } from '../../shared/components/Logo'
-import { useUserLogout } from '../../entities/user/use-user-logout'
+import { Logo } from '../../../shared/ui'
+import { useUserLogout } from '../../../entities/user/use-user-logout'
 import { LogoutOutlined } from '@ant-design/icons'
 import './protected-routes.pcss'
