+++ packages/client/src/app/ui/routing/router.tsx
@@ -1,17 +1,17 @@
 import { createBrowserRouter } from 'react-router-dom'
-import { HomePage } from '../pages/home'
-import { ROUTES } from '../shared/config'
-import { ForumPage } from '../pages/forum'
-import { ForumTopicPage } from '../pages/forum-topic'
-import { GamePage } from '../pages/game'
-import { LeaderboardPage } from '../pages/leaderboard'
-import { ProfilePage } from '../pages/profile'
-import { SignInPage } from '../pages/sign-in'
-import { SignUpPage } from '../pages/sign-up'
-import { ProtectedRoutes } from './ui/protected-routes'
-import { NoAuthRoutes } from './ui/no-auth-routes'
-import { ErrorElement } from './ui/error-element'
-import { ErrorPage } from '../pages/error-page'
+import { HomePage } from '../../../pages/home'
+import { ROUTES } from '../../../shared/config'
+import { ForumPage } from '../../../pages/forum'
+import { ForumTopicPage } from '../../../pages/forum-topic'
+import { GamePage } from '../../../pages/game'
+import { LeaderboardPage } from '../../../pages/leaderboard'
+import { ProfilePage } from '../../../pages/profile'
+import { SignInPage } from '../../../pages/sign-in'
+import { SignUpPage } from '../../../pages/sign-up'
+import { ProtectedRoutes } from './protected-routes'
+import { NoAuthRoutes } from './no-auth-routes'
+import { ErrorElement } from '../error-element'
+import { ErrorPage } from '../../../pages/error-page'
 
 export const router = createBrowserRouter([
   {
