+++ packages/client/src/pages/error-page/error-page.module.pcss
@@ -1,11 +1,3 @@
-.layout {
-  height: 100vh;
-
-  .container {
-    height: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    flex-direction: column;
-  }
+.container {
+  height: 100%;
 }
