--- packages/client/src/shared/ui/common/router-link.tsx
@@ -1,21 +1,0 @@
-import { Typography } from 'antd'
-import { Link } from 'react-router-dom'
-import { ReactNode } from 'react'
-
-const { Link: AntLink } = Typography
-
-export function RouterLink({
-  to,
-  children,
-}: {
-  to: string
-  children: ReactNode
-}) {
-  return (
-    <AntLink>
-      <span>
-        <Link to={to}>{children}</Link>
-      </span>
-    </AntLink>
-  )
-}
