+++ packages/client/src/shared/ui/index.ts
@@ -1,10 +1,10 @@
-export { AuthLayout } from './layout/auth-layout'
-export { AuthForm } from './auth-form/auth-form'
+export { AuthForm } from './auth-form'
 export { LoginField } from './fields/login-field'
 export { FirstNameField } from './fields/first-name-field'
 export { SecondNameField } from './fields/second-name-field'
 export { EmailField } from './fields/email-field'
 export { PasswordField } from './fields/password-field'
 export { PhoneField } from './fields/phone-field'
-export { RouterLink } from './common/router-link'
 export { useError } from './use-error'
+export { AppLayout } from './app-layout'
+export { Logo } from './Logo'
