+++ packages/client/src/shared/ui/logo.module.pcss
@@ -1,0 +1,8 @@
+.link {
+  max-height: 70px;
+
+  .logo {
+    width: 196px;
+    height: 70px;
+  }
+}
